Extract class responsible for tinygo info parsing
Extract assembly of tinygo command line arguments and parsing of the
output into TinyGoInfoExtractor
Moved tinygo settings to go subfolder
