+++ src/main/kotlin/org/jetbrains/tinygoplugin/services/TinyGoSettingsService.kt
@@ -1,6 +1,5 @@
 package org.jetbrains.tinygoplugin.services
 
-import com.goide.util.GoExecutor
 import com.goide.util.GoHistoryProcessListener
 import com.intellij.openapi.diagnostic.Logger
 import com.intellij.openapi.options.Configurable
@@ -18,6 +17,7 @@ class TinyGoSettingsService(private val project: Project) : Configurable, Action
     }
 
     lateinit var settingsUI: TinyGoSettingsUI
+    val infoExtractor = TinyGoInfoExtractor(project)
 
     private fun setSettingsToUI(settings: TinyGoConfiguration) {
         settingsUI.targetPlatform = settings.targetPlatform
@@ -57,55 +57,20 @@ class TinyGoSettingsService(private val project: Project) : Configurable, Action
     override fun actionPerformed(e: ActionEvent?) {
         logger.debug("Starting tinygo tags detection")
         settingsUI.setDetectionInProgress()
-        val executor = GoExecutor.`in`(project, null)
-        val settings = TinyGoConfiguration.getInstance(project)
-        /* ktlint-disable*/
-        val parameters = listOf(
-            "-target", settings.targetPlatform,
-            "-scheduler", settings.scheduler.cmd,
-            "-gc", settings.gc.cmd
-        )
-            .joinToString(" ", " ", " ")
-        /* ktlint-enable */
-        executor.withParameterString("info $parameters")
-        executor.showNotifications(true, false)
-        val tinyGoExec = Paths.get(
-            Paths.get(settings.tinyGoSDKPath).toAbsolutePath().toString(),
-            "bin",
-            "tinygo"
-        )
-        logger.debug("Tinygo path: $tinyGoExec")
-        logger.debug("Tinygo parameters: $parameters")
-        executor.withExePath(tinyGoExec.toString())
+        // Apply settings from ui to the model
+        apply()
+        val executor = infoExtractor.assembleTinyGoShellCommand()
         val processHistory = GoHistoryProcessListener()
         executor.executeWithProgress(true, true, processHistory, null) { result ->
-            logger.warn("${result.status}")
-            logger.warn("${result.message}")
             val output = processHistory.output.joinToString("")
-            logger.warn(output)
-            extractTinyGoInfo(output)
+            logger.trace(output)
+            infoExtractor.extractTinyGoInfo(output)
+
+            val settings = TinyGoConfiguration.getInstance(project)
             settingsUI.goArch = settings.goArch
             settingsUI.goOS = settings.goOS
             settingsUI.goTags = settings.goTags
             settingsUI.onProcessingFinished()
         }
-    }
-
-    private fun extractTinyGoInfo(msg: String) {
-        val tagPattern = Regex("""build tags:\s+(.+)\n""")
-        val goArchPattern = Regex("""GOARCH:\s+(.+)\n""")
-        val goOSPattern = Regex("""GOOS:\s+(.+)\n""")
-
-        val tags = tagPattern.findAll(msg).first()
-        val goArch = goArchPattern.findAll(msg).first()
-        val goOS = goOSPattern.findAll(msg).first()
-
-        val settings = TinyGoConfiguration.getInstance(project)
-
-        settings.goArch = goArch.groupValues[1]
-        settings.goTags = tags.groupValues[1]
-        settings.goOS = goOS.groupValues[1]
-
-        logger.warn("extraction finished")
     }
 }
