+++ src/main/resources/META-INF/plugin.xml
@@ -12,7 +12,7 @@
 
     <extensions defaultExtensionNs="com.intellij">
 
-        <projectConfigurable parentId="tools" instance="org.jetbrains.tinygoplugin.services.TinyGoSettingsService"
+        <projectConfigurable parentId="go" instance="org.jetbrains.tinygoplugin.services.TinyGoSettingsService"
                              id="org.jetbrains.tinygoplugin.services.TinyGoSettingsService"
                              displayName="TinyGo"/>
     </extensions>
